requirements.txt
pyTelegramBotAPI
telebot
google.oauth2.credentials
googleapiclient.discovery
